% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordfeatures-package.R
\docType{package}
\name{wordfeatures-package}
\alias{wordfeatures}
\alias{wordfeatures-package}
\title{wordfeatures: Lexical Feature Data and Embedding Helpers}
\description{
Package-level documentation for `wordfeatures`.
}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
